+++ modules/core/src/index.js
@@ -36,7 +36,6 @@ export {toArrayBuffer} from './javascript-utils/binary-utils';
 // ITERATOR UTILS
 export {makeIterator} from './iterator-utils/make-iterator/make-iterator';
 
-export {forEach, concatenateChunksAsync} from './iterator-utils/async-iteration';
 export {
   makeTextDecoderIterator,
   makeTextEncoderIterator,
@@ -49,6 +48,7 @@ export {isBrowser, isWorker, self, window, global, document} from '@loaders.gl/l
 export {assert} from '@loaders.gl/loader-utils';
 export {setPathPrefix, getPathPrefix, resolvePath} from '@loaders.gl/loader-utils';
 export {RequestScheduler} from '@loaders.gl/loader-utils';
+export {forEach, concatenateChunksAsync} from '@loaders.gl/loader-utils';
 
 // EXPERIMENTAL
 export {default as _WorkerThread} from './worker-utils/worker-thread';
