--- modules/core/src/javascript-utils/memory-copy-utils.d.ts
@@ -1,4 +1,0 @@
-export function concatenateArrayBuffers(
-  source1: ArrayBuffer | Uint8Array,
-  source2: ArrayBuffer | Uint8Array
-): ArrayBuffer;
