+++ modules/loader-utils/src/lib/binary-utils/memory-copy-utils.d.ts
@@ -1,16 +1,27 @@
 /**
- * 
- * @param byteLength 
+ *
+ * @param byteLength
  */
 export function padTo4Bytes(byteLength);
 
 /**
  * Copy a view of an ArrayBuffer into new ArrayBuffer with byteOffset = 0
- * @param arrayBuffer 
- * @param byteOffset 
- * @param byteLength 
+ * @param arrayBuffer
+ * @param byteOffset
+ * @param byteLength
  */
 export function getZeroOffsetArrayBuffer(arrayBuffer, byteOffset, byteLength);
+
+/**
+ * Concatenate two ArrayBuffers
+ * @param source1 The first ArrayBuffer.
+ * @param source2 The second ArrayBuffer.
+ * @return A concatenated ArrayBuffer
+ */
+export function concatenateArrayBuffers(
+  source1: ArrayBuffer | Uint8Array,
+  source2: ArrayBuffer | Uint8Array
+): ArrayBuffer;
 
 /**
  * Creates a new Uint8Array based on two different ArrayBuffers
@@ -34,4 +45,4 @@ export function copyArrayBuffer(
  *
  * @return the new offset taking into account proper padding
  */
-export function copyToArray(source: ArrayBuffer|any, target: any, targetOffset: number): number;
+export function copyToArray(source: ArrayBuffer | any, target: any, targetOffset: number): number;
