+++ modules/loader-utils/src/lib/binary-utils/memory-copy-utils.js
@@ -11,6 +11,16 @@ export function getZeroOffsetArrayBuffer(arrayBuffer, byteOffset, byteLength) {
   return arrayCopy.buffer;
 }
 
+// Concatenate two ArrayBuffers
+export function concatenateArrayBuffers(source1, source2) {
+  const sourceArray1 = source1 instanceof ArrayBuffer ? new Uint8Array(source1) : source1;
+  const sourceArray2 = source2 instanceof ArrayBuffer ? new Uint8Array(source2) : source2;
+  const temp = new Uint8Array(sourceArray1.byteLength + sourceArray2.byteLength);
+  temp.set(sourceArray1, 0);
+  temp.set(sourceArray2, sourceArray1.byteLength);
+  return temp.buffer;
+}
+
 /* Creates a new Uint8Array based on two different ArrayBuffers
  * @private
  * @param {ArrayBuffers} buffer1 The first buffer.
